10/requirements.txt
tweepy==3.5.0
Flask==0.12.2
nltk==3.2.5
pymongo==3.4.0
Flask_Cors==3.0.2
pytz==2016.10
python_twitter==3.3
